fix curse and flying machines & move function to anticheese
