+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/curse/CurseAscensionListener.kt
@@ -8,10 +8,10 @@ import com.mineinabyss.idofront.messaging.broadcast
 import com.mineinabyss.mineinabyss.core.AbyssContext
 import com.mineinabyss.mineinabyss.core.isAbyssWorld
 import com.mineinabyss.mineinabyss.core.layer
-import dev.geco.gsit.api.GSitAPI
 import dev.geco.gsit.api.event.PlayerGetUpSitEvent
 import io.papermc.paper.event.entity.EntityMoveEvent
 import org.bukkit.Location
+import org.bukkit.block.BlockFace
 import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
@@ -59,19 +59,25 @@ class CurseAscensionListener : Listener {
             handleCurse(player, from, to)
     }
 
+    @EventHandler(ignoreCancelled = true)
+    fun BlockPistonExtendEvent.handleCurseOnFlyingMachine() {
+        if (direction != BlockFace.UP) return
+        blocks.forEach {
+            val players = it.location.getNearbyPlayers(1.0, 1.0, 1.0)
+
+            players.forEach { player ->
+                val from = player.location
+                val to = player.location.apply { y += 1 }
+                handleCurse(player, from, to)
+                broadcast(player.playerData.curseAccrued)
+            }
+        }
+    }
+
     @EventHandler
     fun PlayerGetUpSitEvent.handleCurseOnSitting() {
         if (!AbyssContext.isGSitLoaded) return
         handleCurse(player, seat.location.toBlockLocation(), player.location)
-        broadcast(player.playerData.curseAccrued)
-    }
-
-    // Cancels pistons if a player is riding it via a GSit Seat
-    @EventHandler
-    fun BlockPistonExtendEvent.seatMovedByPiston() {
-        if (!AbyssContext.isGSitLoaded) return
-        if (GSitAPI.getSeats(blocks).isEmpty()) return
-        else isCancelled = true
     }
 
     private fun handleCurse(player: Player, from: Location, to: Location) {
